+++ plugin_ide.ui/src/melnorme/lang/ide/ui/OperationsConsolePrefDefaults_Actual.java
@@ -1,0 +1,22 @@
+/*******************************************************************************
+ * Copyright (c) 2014, 2014 Bruno Medeiros and other Contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     Bruno Medeiros - initial API and implementation
+ *******************************************************************************/
+package melnorme.lang.ide.ui;
+
+import org.eclipse.swt.graphics.RGB;
+
+public interface OperationsConsolePrefDefaults_Actual {
+	
+	static RGB INFO_COLOR_DEFAULT = new RGB(0, 0, 200);
+	static RGB STDERR_COLOR_DEFAULT = new RGB(255, 0, 0);
+	static RGB STDOUT_COLOR_DEFAULT = new RGB(0, 0, 0);
+	static RGB BACKGROUND_COLOR_DEFAULT = new RGB(124, 232, 52); // Change this: LANGUAGE_
+	
+}