+++ plugin_tooling/src/dtool/ast/declarations/AbstractConditionalDeclaration.java
@@ -4,7 +4,7 @@ import static melnorme.utilbox.core.Assert.AssertNamespace.assertTrue;
 
 import java.util.Iterator;
 
-import melnorme.utilbox.misc.ChainedIterator;
+import melnorme.utilbox.collections.ChainedIterator;
 import melnorme.utilbox.misc.IteratorUtil;
 import dtool.ast.ASTCodePrinter;
 import dtool.ast.ASTNode;